*** empty log message ***
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/gworkspace/trunk@19177 72102866-910b-0410-8b05-ffd578937521

+++ FSNode/FSNIcon.h
@@ -81,7 +81,7 @@
 }
 
 - (id)initForNode:(FSNode *)anode
-         iconSize:(float)isize
+         iconSize:(int)isize
      iconPosition:(unsigned int)ipos
         gridIndex:(int)gindex
         dndSource:(BOOL)dndsrc
